fix : 깨지던 테스트 수정
